safari-600.7-branch/LayoutTests/js/map-repack-with-object-keys-expected.txt

Tests to make sure we correctly repack a Map with object keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Array.isArray(map.get(newObject1)) is true
PASS Array.isArray(map.get(newObject1)) is true
PASS successfullyParsed is true

TEST COMPLETE
